Added new STABILIZATION state
I noticed that after "bad" frames, the calibration would often be
flooded with additional noisy frames. So, I've added a stabilization
state, during where the L1 process just drops a fixed number of frames
before transitioning to CALIBRATION.

This state is now the *only* valid pathway to CALIBRATION in the FSM. It
is also the first state on startup (often there is some noise when the
user is first starting the app, i.e. before it's set flat on a table).

I also changed the output logic so that aborted calibration blocks are
no longer uploaded.