Reworked ExposureBlock management in DAQActivity
Added a new class, `ExposureBlockManager`, which provides thread-safe
access to exposure blocks for the DAQ logic.
The current XB is also bundled with the raw frame data at acquisition
time, so there's no need to try to re-figure which frame goes to which
XB.
The method for committing old XB's should also be more robust now; it
waits until the closing time of the XB is earlier than the latest-seen
frame, before concluding it can commit the data.