Fix issues with resuming DAQ after app suspended
There are some nuances in dealing with the lifetime of the `Camera` object
and the `CameraPreview` object. It turns out that the `CameraPreview` is not
destroyed and/or modified every time `onPause()` is called, which is why
the preview data was not resuming in some cases. It should now resume in
every case.

Also added `IDLE` state, which the app transitions to at `onPause()`.
This is to cleaning end the current ExposureBlock, so that we don't count
times that the app is suspended towards our exposure calculations.