Added support for local filestore caching
The OutputManager will first check if there is a network connection
available. If not, it will dump the protobuf binary to a local file.

The main loop of OutputManager checks for local binary files once per
loop (if the network is available), and if it finds any, then they are
uploaded to the server.