Fix calibration state transition bug.
There was a bug in doStateTransition: the switch() was on the current
state, rather than the previous state. This was preventing the necessary
setup for calibration to be run, but the app was still transitioning
it's state.

We should fail harder if there is an illegal state transition: the whole
point of the FSM is that it needs to be airtight.