Oops, merged the wrong FSM handling code (I think)
Originally I had merged in the new FSM exceptions, but kept the part
where all exceptions were being caught. We decided earlier that
these FSM exceptions should be fatal (and of course must be fixed,
if they are occuring).

This also negates the previous commit (the Crashlytics import was
only needed by the exception handling code, to report the non-fatal).