Add resolution info to exposureblocks
The protobuf definitions have been updated to carry information about
resolution for each XB. In the future we may have dynamic resolution
settings, but it will always be constant on the XB interval.

I also simplified some areas of the code where the camera resolution was
being passed around all over the place. As of a few months ago, the
RawCameraFrame object keeps track of its own resolution, so all cases
where the previewSize was being used before can be inferred by
inspecting the frame.