Do not truncate buffer length to match dimensions!
It's not a good idea to truncate the raw image buffer length, in the
case that it doesn't match up with the expected image dimensions. Since
we only read the first ~1/3 of the buffer (Y channel of the YUV format),
we would end up reading in garbage and making meaningless comparisons
between different channels. It also breaks the index->(x,y) mapping which
depends on the image width. The result is that the app happily gives us
garbage data, and we'd never know about it.

It's much better to simply let the app throw an exception if the buffer
is the wrong size. This way (if it ever happens) we can be alerted via
Crashlytics and fix the underlying problem.