Tons of changes! Modular L2 (no pixel reco, yet)
This is a huge commit; so many things were entangled that I couldn't
break it down into smaller refactorizations.

The main goal was to modularize the L2 processing logic. This is pretty
much accomplished (although I haven't yet re-implemented the pixel
selection algorithm in the new framework). More details on this later,
once it is complete. But the basic idea is to do away with `L2thread` and
`PARTICLE_RECO`, and instead make a hot-swappable `L2Task` interface that
is dispatched whenever a frame is accepted at L1.

The major side effects of this commit include:
 - Some changes to the XB flush logic; it now runs in a dedicated BG thread
   at a fixed delay after each new XB is created.
 - XB's operate on nanotime now.
 - Fix threshold (aka trigger lock) is now an option that can be
   accessed through CFConfig (and hence should be easy to set remotely)
 - As a side effect of the above, now all preferences are written to the
   same location. before UI preferences and all other preferences were
   saved independently, which makes it hard to access both from CFConfig.
   not 100% if this is going to cause problems, but it seems okay sofar.
 - Broke some UI elements that were getting their info from the old L2
   apparatus. These can be patched up later.