Add L2 pixel selection logic. Also upgrade event-lvl bg stats.
L2 pixel selection should be working as before.
Now when events are selected, the BG average level is computed with the
full statistics over all pixels (since we have to loop anyways), and the
event is updated with this value.

I also started keeping track of the running average bg levels over all
frames observed by an XB, although this does not get written to the
protobuf output (yet).