Better bg statistics
Just realized we can do even better with the BG statistics. We get the
full-sensor average for free on _every_ frame since we have to look for the maximum
anyways. Then since we know the average, we can do only one "quick" pass
to calculate the variance in `buildEvent()`. Finally, for events accepted
by L1, we have to traverse all pixels again, so we can recalculate the
full-sensor variance and update the `RecoEvent` object for frames selected
by the L1 trigger.